+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Neutral/LivingFlame.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using HearthDb.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Neutral;
+
+public class LivingFlame : ICardWithHighlight
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Neutral.LivingFlame;
+
+	public HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck) =>
+		HighlightColorHelper.GetHighlightColor(card.GetTag(GameTag.SPELL_SCHOOL) == (int)SpellSchool.FIRE);
+}